Refactoring data model for UI

+++ surf/Model/Snapshot.swift
@@ -35,7 +35,7 @@ struct Snapshot {
     var currentTideDirection: String?
     var stationName : String?
     var stationId : Int?
-    var beachFaceDirection : Double?
+    var beachFaceDirection : Int?
     var windDir: Double?
     var windSpd: Double?
     var windDirectionString: String?
